Gave website a template
Created the basic look for a website using BootStrap 5. Created a navigation bar, and footer. I also get rid of the current form and updated it with one that has all the necessary variables to be able to be plugged into some javascript code to do the calculations.

I left the rest of the site empty as it basically just needs a descript as to what the formula does, and explanation of it.